// frontend/src/app/core/api/generated/model/jobLogList.ts
/**
 * Plataforma Científica Modular API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { JobLogEvent } from './jobLogEvent';


/**
 * Contrato de respuesta para listado paginado de logs por job.
 */
export interface JobLogList { 
    /**
     * Identificador único del job.
     */
    job_id: string;
    /**
     * Cantidad de eventos en la página de resultados.
     */
    count: number;
    /**
     * Cursor recomendado para consultar la siguiente página.
     */
    next_after_event_index: number;
    results: Array<JobLogEvent>;
}