// frontend/src/app/core/api/generated/model/progressStageEnum.ts
/**
 * Chemistry Apps API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `pending` - pending * `queued` - queued * `running` - running * `paused` - paused * `recovering` - recovering * `caching` - caching * `completed` - completed * `failed` - failed * `cancelled` - cancelled
 */
export const ProgressStageEnum = {
    Pending: 'pending',
    Queued: 'queued',
    Running: 'running',
    Paused: 'paused',
    Recovering: 'recovering',
    Caching: 'caching',
    Completed: 'completed',
    Failed: 'failed',
    Cancelled: 'cancelled'
} as const;
export type ProgressStageEnum = typeof ProgressStageEnum[keyof typeof ProgressStageEnum];